# [emval] Reuse method calling optimisation in regular calls