--- a/src/embind/emval.js
+++ b/src/embind/emval.js
@@ -213,14 +213,22 @@ var LibraryEmVal = {
     handle[key] = value;
   },
 
-  _emval_as__deps: ['$Emval', '$requireRegisteredType'],
+  $emval_returnValue__deps: ['$Emval'],
+  $emval_returnValue: (returnType, destructorsRef, handle) => {
+    var destructors = [];
+    var result = returnType['toWireType'](destructors, handle);
+    if (destructors.length) {
+      // void, primitives and any other types w/o destructors don't need to allocate a handle
+      {{{ makeSetValue('destructorsRef', '0', 'Emval.toHandle(destructors)', '*') }}};
+    }
+    return result;
+  },
+
+  _emval_as__deps: ['$Emval', '$requireRegisteredType', '$emval_returnValue'],
   _emval_as: (handle, returnType, destructorsRef) => {
     handle = Emval.toValue(handle);
     returnType = requireRegisteredType(returnType, 'emval::as');
-    var destructors = [];
-    var rd = Emval.toHandle(destructors);
-    {{{ makeSetValue('destructorsRef', '0', 'rd', '*') }}};
-    return returnType['toWireType'](destructors, handle);
+    return emval_returnValue(returnType, destructorsRef, handle);
   },
 
   _emval_as_int64__deps: ['$Emval', '$requireRegisteredType'],
@@ -310,9 +318,9 @@ var LibraryEmVal = {
   `,
 
   _emval_get_method_caller__deps: [
-    '$emval_addMethodCaller', '$emval_lookupTypes',,
+    '$emval_addMethodCaller', '$emval_lookupTypes',
     '$makeLegalFunctionName',
-    '$reflectConstruct',
+    '$reflectConstruct', '$emval_returnValue',
 #if DYNAMIC_EXECUTION
     '$newFunc',
 #endif
@@ -324,7 +332,7 @@ var LibraryEmVal = {
 
 #if !DYNAMIC_EXECUTION
     var argN = new Array(argCount);
-    var invokerFunction = (obj, func, destructors, args) => {
+    var invokerFunction = (obj, func, destructorsRef, args) => {
       var offset = 0;
       for (var i = 0; i < argCount; ++i) {
         argN[i] = types[i]['readValueFromPointer'](args + offset);
@@ -336,13 +344,13 @@ var LibraryEmVal = {
           types[i].deleteObject(argN[i]);
         }
       }
-      return retType['toWireType'](destructors, rv);
+      return emval_returnValue(retType, destructorsRef, rv);
     };
 #else
     var signatureName = retType.name + "_$" + types.map(t => t.name).join("_") + "$";
     var functionName = makeLegalFunctionName("methodCaller_" + signatureName);
     var functionBody =
-        "return function " + functionName + "(obj, func, destructors, args) {\n";
+        "return function " + functionName + "(obj, func, destructorsRef, args) {\n";
 
     var offset = 0;
     var argsList = asCtor ? "" : "obj"; // 'obj?, arg0, arg1, arg2, ... , argN'
@@ -366,8 +374,10 @@ var LibraryEmVal = {
         }
     }
     if (!retType.isVoid) {
-        functionBody +=
-        "    return retType.toWireType(destructors, rv);\n";
+      params.push("emval_returnValue");
+      args.push(emval_returnValue);
+      functionBody +=
+      "    return emval_returnValue(retType, destructorsRef, rv);\n";
     }
     functionBody +=
         "};\n";
@@ -381,13 +391,7 @@ var LibraryEmVal = {
   $callAndDestruct__deps: ['$emval_methodCallers'],
   $callAndDestruct: (caller, obj, func, destructorsRef, args) => {
     caller = emval_methodCallers[caller];
-    var destructors = [];
-    var result = caller(obj, func, destructors, args);
-    // void and any other types w/o destructors don't need to allocate a handle
-    if (destructors.length) {
-      {{{ makeSetValue('destructorsRef', '0', 'Emval.toHandle(destructors)', '*') }}};
-    }
-    return result;
+    return caller(obj, func, destructorsRef, args);
   },
 
   _emval_call_method__deps: ['$getStringOrSymbol', '$callAndDestruct', '$Emval'],