… src/library_sigs.js