Method, function and constructor calls are very similar in nature but used different code paths: method calls have an optimisation where an optimised caller trampoline is generated once and reused for each function signature, whereas the other two serialized arguments and types every time.

There is no reason for them to diverge, so I made function calls and constructors reuse the same optimisation.

Results on a silly microbenchmark running a side-effect-ful JS function with 100 integer args in a loop:

### Before

#### with dynamic execution:

as method: 224.504ns
as function: 5367.9ns

#### without dynamic execution:

as method: 845.044ns
as function: 5350.96ns

### After

#### with dynamic execution:

as method: 262.74ns
as function: 246.593ns

#### without dynamic execution:

as method: 859.534ns
as function: 823.498ns

In real world the difference is unlikely to be perceptible, but at the very least this makes the code more consistent and deletes a bunch of duplication.