@RReverser 2023-10-07
FWIW I'm happy to go in the opposite direction too and remove the optimisation if we decide that complexity of maintaining it is not worth it given that speed is measured in microseconds even for complex functions.

I just want those callers to converge on single codepath before I make more PRs tweaking that shared codepath.